[clean] Moved Marianas homework in a new package
