--- file.txt
@@ -1,1 +1,0 @@
-First pull request.
