--- mihai.txt
@@ -1,2 +1,0 @@
-hej
-Added a description for u
