+++ src/com/java/training/d03/homework/mariana/account/AbstractAccount.java
@@ -1,0 +1,4 @@
+package com.java.training.d03.homework.mariana.account;
+
+public class AbstractAccount {
+}
