+++ src/com/java/training/d03/homework/mariana/account/Admin.java
@@ -1,0 +1,4 @@
+package com.java.training.d03.homework.mariana.account;
+
+public class Admin {
+}
