+++ src/com/java/training/d03/homework/mariana/car/interf/CarRepository.java
@@ -1,0 +1,9 @@
+package com.java.training.d03.homework.mariana.car.interf;
+
+import com.java.training.d03.homework.mariana.car.types.AbstractCar;
+
+public interface CarRepository {
+
+    AbstractCar getCar ();
+
+}
