+++ src/com/java/training/d03/homework/mariana/car/types/ServiceCar.java
@@ -1,0 +1,8 @@
+package com.java.training.d03.homework.mariana.car.types;
+
+public class ServiceCar {
+
+    public void changeColor(AbstractCar car, String color){
+        car.setColor(color);
+    }
+}
