+++ src/com/java/training/d03/homework/mariana/fruits/Apple.java
@@ -1,0 +1,4 @@
+package com.java.training.d03.homework.mariana.fruits;
+
+public class Apple {
+}
