+++ src/com/java/training/d03/homework/mariana/settlement/AbstractSettlement.java
@@ -1,0 +1,4 @@
+package com.java.training.d03.homework.mariana.settlement;
+
+public class AbstractSettlement {
+}
