+++ src/com/java/training/d03/homework/mariana/settlement/Town.java
@@ -1,0 +1,4 @@
+package com.java.training.d03.homework.mariana.settlement;
+
+public class Town {
+}
