--- src/com/java/training/mariana/sesiune3/account/AbstractAccount.java
@@ -1,4 +1,0 @@
-package com.java.training.mariana.sesiune3;
-
-public class AbstractAccount {
-}
