--- src/com/java/training/mariana/sesiune3/account/Admin.java
@@ -1,4 +1,0 @@
-package com.java.training.mariana.sesiune3.account;
-
-public class Admin {
-}
