--- src/com/java/training/mariana/sesiune3/account/User.java
@@ -1,4 +1,0 @@
-package com.java.training.mariana.sesiune3.account;
-
-public class User {
-}
