--- src/com/java/training/mariana/sesiune3/car/interf/CarRepository.java
@@ -1,9 +1,0 @@
-package com.java.training.mariana.sesiune3.car.interf;
-
-import com.java.training.mariana.sesiune3.car.types.AbstractCar;
-
-public interface CarRepository {
-
-    AbstractCar getCar ();
-
-}
