--- src/com/java/training/mariana/sesiune3/car/types/ServiceCar.java
@@ -1,8 +1,0 @@
-package com.java.training.mariana.sesiune3.car.types;
-
-public class ServiceCar {
-
-    public void changeColor(AbstractCar car, String color){
-        car.setColor(color);
-    }
-}
