--- src/com/java/training/mariana/sesiune3/fruits/AbstractFruit.java
@@ -1,4 +1,0 @@
-package com.java.training.mariana.sesiune3.fruits;
-
-public class AbstractFruit {
-}
