--- src/com/java/training/mariana/sesiune3/fruits/Apple.java
@@ -1,4 +1,0 @@
-package com.java.training.mariana.sesiune3.fruits;
-
-public class Apple {
-}
