--- src/com/java/training/mariana/sesiune3/fruits/Strawberry.java
@@ -1,4 +1,0 @@
-package com.java.training.mariana.sesiune3.fruits;
-
-public class Strawberry {
-}
