--- src/com/java/training/mariana/sesiune3/settlement/AbstractSettlement.java
@@ -1,4 +1,0 @@
-package com.java.training.mariana.sesiune3.settlement;
-
-public class AbstractSettlement {
-}
