--- src/com/java/training/mariana/sesiune3/settlement/Town.java
@@ -1,4 +1,0 @@
-package com.java.training.mariana.sesiune3.settlement;
-
-public class Town {
-}
